adds the hierarchy and cmake, fixes #8 and #11
