+++ etc/styles/defaults.qss
