+++ src/commands/CMakeLists.txt
@@ -1,0 +1,3 @@
+project (commands)
+file(GLOB SOURCES "*.cpp")
+add_library (commands STATIC ${SOURCES})
