+++ src/core/CMakeLists.txt
@@ -1,0 +1,3 @@
+project (core)
+file(GLOB SOURCES "*.cpp")
+add_library (core STATIC ${SOURCES})
