+++ src/core/runtime_environment.hpp
