+++ src/core/working_set.cpp
