+++ src/core/working_set.hpp
