+++ src/gui/qribbon.cpp
@@ -1,14 +1,17 @@
 #include "qribbon.hpp"
 
+#include "icons.hpp"
+
 #include <QLayout>
 #include <QFile>
+
 
 QRibbon::QRibbon(QWidget* p)
 	: QTabWidget(p)
 {
 	setFixedHeight(127);
 
-	QFile styles("C:\\Users\\elen\\Desktop\\my_projects\\qt_painter\\src\\painter\\defaults.qss");
+	QFile styles(getStylesDir()+"defaults.qss");
 	styles.open(QFile::ReadOnly);
 	QString styleSheet = QLatin1String(styles.readAll());
 	setStyleSheet(styleSheet);
