+++ src/gui/qribbon.hpp
