+++ src/gui/qribbonbutton.cpp
