+++ src/gui/qribbongroup.cpp
