+++ src/gui/qribbongroup.hpp
