+++ src/gui/qribbonwidget.cpp
