+++ src/gui/qribbonwidget.hpp
