+++ src/io/CMakeLists.txt
@@ -1,0 +1,3 @@
+project (io)
+file(GLOB SOURCES "*.cpp")
+add_library (io STATIC ${SOURCES})
